requirements.txt
streamlit
numpy
openai-whisper
ffmpeg-python
